rclgo-gen: Autodetect message-module-prefix
When the generator is run inside of a go module it'll read the full
package path and use that as the message-module-prefix. It'll fall back to
the default value specified if the package name check fails.

destPath can be used to modify the prefix if the generated output
differs from the current directory the generator is running.